requirements.txt
flask
moviepy==1.0.3
decorator==4.4.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
Pillow
numpy<2.0.0
requests
edge-tts
gtts